Adds [+] button animation
